+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## 6.x
+
+Changes:
+
+- Check `@polkadot/hw-ledger` against `@polkadot/networks` for discrepancies (test-only)
+
 ## 6.1.1 Apr 11, 2021
 
 Contributed: